requirements.txt
emoji==2.7.0
mido==1.3.0
packaging==23.1
python-dotenv==1.0.0
python-rtmidi==1.5.5
